Knapsack update 2 (read more)
Knapsack now supports importing solutions
Instead of checking whenever fitness is zero, we allow custom solution testing using isSolution()
MaxIteration changed to 1000 from 16384
In SUS, changed arregateInvertFitness to long (if the fitness is to high, overflows the int)

+++ src/com/AILab3/GeneticAlgo/Selection.java
@@ -26,7 +26,7 @@ public class Selection
                                                     boolean aging)
     {
         parentsRoulette.clear();
-        int[] aggregateInvertFitness = new int[GA_POPSIZE];
+        long[] aggregateInvertFitness = new long[GA_POPSIZE];
         int first = 0;
         if (aging)
             // We don't want the children to reproduce, children have age of 1 now
@@ -40,7 +40,7 @@ public class Selection
         // distance between the drawn fitness values
         // the last element of aggregate function will always contain the total sum
         int rouletteSize = aging ? amountOfChildrenToGenerate * 2 : GA_POPSIZE / 4;
-        int fitnessJump = (aggregateInvertFitness[GA_POPSIZE - 1] / rouletteSize);
+        int fitnessJump = (int) (aggregateInvertFitness[GA_POPSIZE - 1] / rouletteSize);
         // start randomly between 0 and the jump distance
         int start = r.nextInt(fitnessJump);
         int populationIterator = first;
